[layouts] Don't tile windows to size -1
If a window had a size < the size of the border, when adjusting the size
for the border, the target size would go negative. However, the variable
was unsigned so it wrapped to max short value so the default min
width/height wasn't applied. Now values are casted to (signed) int when
checking if the min value of 1 should be applied or not.

+++ src/layouts.cpp
@@ -203,8 +203,8 @@ void configureWindow(const LayoutState* state, const WindowInfo* winInfo, const 
             else
                 config[i] -= (&state->getArgs()->rightPadding)[i % 2] + (&state->getArgs()->leftPadding)[i % 2];
         }
-    config[CONFIG_INDEX_WIDTH] = MAX(1, config[CONFIG_INDEX_WIDTH]);
-    config[CONFIG_INDEX_HEIGHT] = MAX(1, config[CONFIG_INDEX_HEIGHT]);
+    config[CONFIG_INDEX_WIDTH] = MAX(1, (short)config[CONFIG_INDEX_WIDTH]);
+    config[CONFIG_INDEX_HEIGHT] = MAX(1, (short)config[CONFIG_INDEX_HEIGHT]);
     assert(winInfo->getID());
     configureWindow(winInfo->getID(), mask, config);
 }
